Don't use a Modded field + use specific file instead of Everest module save data
